some variables introduced + .desktop fixed

+++ minimize/home/devuan/Desktop/minimize/doIt6.sh
@@ -1,26 +1,42 @@
 #!/bin/sh
 
+iface=eth0
+debug=1
+the_ar=0
+tblz4=rules.v4.180624
+install=0 #debug ja install komentoriviparam kautta jatkossa
+tgtfile=/mnt/186/nu.tar
+
 #ch-jutut siltä varalta että tar sössii oikeudet tai omistajat
 sudo chown root:root /home
 sudo chmod 0755 /home
 sudo chown -R devuan:devuan /home/devuan/
 #sudo chown devuan:devuan /home/devuan/Desktop/
 sudo chmod -R 0755 /home/devuan/Desktop/minimize
+sudo chown -R 101:65534 /home/stubby/
 
-#TODO:testaa tästä eteenpäin ch-kikkailut
-sudo chown -R root:root /sbin
-sudo chmod -R 0755 /sbin
+if [ ${debug} -eq 1 ] ; then 
+	#TODO:testaa tästä eteenpäin ch-kikkailut (tilap jemmassa)
+	echo "https://raw.githubusercontent.com/senescent777/project/main/sbin/ can be used as /sbin"
+	echo "sudo chown -R root:root /sbin"
+	echo "sudo chmod -R 0755 /sbin"
 
-sudo chown -R root:root /etc
-sudo chmod -R go-w /etc
-sudo chown -R root:root /var
-sudo chmod -R go-w /var
-sudo chown root:root /
-sudo chmod 0755 /
+	echo "https://raw.githubusercontent.com/senescent777/project/main/etc can be used as /etc except for stubby-related stuff"
+	echo "sudo chown -R root:root /etc"
+	echo "sudo chmod -R go-w /etc"
 
-man date #jos tar nalkuttaa päiväyksistä niin date --set hoitaa
-sudo ip link set eth0 down
+	echo "sudo chown -R root:root /var"
+	echo "sudo chmod -R go-w /var"
+	echo "sudo chown root:root /"
+	echo "sudo chmod 0755 /"
+else
+	echo "changing /sbin , /etc and /var 4 real"
+fi
+
+echo "man date; sudo date --set if necessary" #jos tar nalkuttaa päiväyksistä niin date --set hoitaa
+sudo ip link set ${iface} down
 /sbin/ifconfig;sleep 5
+#exit
 
 sudo /usr/sbin/iptables -P INPUT DROP
 sudo /usr/sbin/ip6tables -P INPUT DROP
@@ -33,6 +49,7 @@ sudo /usr/sbin/ip6tables -L;sleep 5
 
 for s in avahi-daemon bluetooth cups cups-browsed exim4 nfs-common network-manager ntp mdadm saned rpcbind lm-sensors ; do
 sudo /etc/init.d/$s stop
+sleep 1
 done
 
 sleep 3 
@@ -40,45 +57,66 @@ sudo /usr/bin/pkill --signal 9 cups*
 sudo /usr/bin/pkill --signal 9 avahi*
 sleep 3
 
-sudo apt-get remove --purge --yes libblu* network*
-sudo apt-get remove --purge --yes libcupsfilters* libgphoto*
+sudo apt-get remove --purge --yes libblu* network* libcupsfilters* libgphoto*
 sudo apt-get remove --purge --yes avahi* blu* cups* exim*
 sudo apt-get remove --purge --yes rpc* nfs* ntp* sntp*
-sudo apt-get remove --purge --yes modem* wireless* wpa* iw
-sudo apt-get remove --purge --yes lm-sensors #mdadm jälkimmäisen poisro saattaa olla huono idea
+sudo apt-get remove --purge --yes modem* wireless* wpa* iw lm-sensors
+#sudo apt-get remove --purge --yes  #mdadm jälkimmäisen poisTo saattaa olla huono idea
 sudo rm -rf /run/live/medium/live/initrd.img*
 sleep 3
 
-#TODO:komentoriviparametrin taakse ar
-sudo apt autoremove --yes
+if [ ${the_ar} -eq 1 ] ; then 
+	sudo apt autoremove --yes
+else
+	echo "autoremove postponed"
+fi
+
 sudo rm -rf /run/live/medium/live/initrd.img*
 sleep 3
-
 sudo netstat -tulpan;sleep 5
 
+
 #Tässä kohtaa mielekästä ajaa tables-komentoja vain jos s.a.autoremove:a EI ajettu.
-echo "sudo /usr/sbin/ip6tables-restore /etc/iptables/rules.v6"
-echo "sudo /usr/sbin/iptables-restore /etc/iptables/rules.v4.180624"
-echo "sudo /usr/sbin/iptables -L;sleep 5"
-echo "sudo /usr/sbin/ip6tables -L;sleep 5"
+if [ ${the_ar} -eq 1 ] ; then 
+	echo "sudo /usr/sbin/ip6tables-restore /etc/iptables/rules.v6"
+	#TODO:lennosta rules.v4 mutilointia, yhdestä sun toisesta projektista mallia
+	echo "sudo /usr/sbin/iptables-restore /etc/iptables/${tblz4}"
+	echo "sudo /usr/sbin/iptables -L;sleep 5"
+	echo "sudo /usr/sbin/ip6tables -L;sleep 5"
+else
+	sudo /usr/sbin/ip6tables-restore /etc/iptables/rules.v6
+	echo "sudo /usr/sbin/iptables-restore /etc/iptables/${tblz4}"
+	sudo /usr/sbin/iptables -L
+	sudo /usr/sbin/iptables -L
+	sleep 5
+fi
 
-#TODO:komentoriviparametrin mukaisella ehdolla kaiutettavien komentojen ajo oikeasti
-echo "sudo /sbin/ifup eth0"
-echo "sudo apt-get update"
-echo "sudo apt-get --no-install-recommends reinstall libip4tc2 libip6tc2 libxtables12 netbase libmnl0 libnetfilter-conntrack3 libnfnetlink0 libnftnl11 iptables"
-sudo rm -rf /run/live/medium/live/initrd.img*
+#exit
 
-echo "sudo apt-get --no-install-recommends reinstall init-system-helpers netfilter-persistent iptables-persistent"
-echo "sudo rm -rf /run/live/medium/live/initrd.img*"
+#VAIH:komentoriviparametrin mukaisella ehdolla kaiutettavien komentojen ajo oikeasti
+if [ ${install} -eq 1 ] ; then 
+	echo "sudo /sbin/ifup ${iface}"
+	echo "sudo apt-get update"
+	echo "sudo apt-get --no-install-recommends reinstall libip4tc2 libip6tc2 libxtables12 netbase libmnl0 libnetfilter-conntrack3 libnfnetlink0 libnftnl11 iptables"
+	sudo rm -rf /run/live/medium/live/initrd.img*
 
-echo "sudo apt-get --no-install-recommends reinstall dnsmasq-base runit-helper"
-echo "sudo rm -rf /run/live/medium/live/initrd.img*"
+	echo "sudo apt-get --no-install-recommends reinstall init-system-helpers netfilter-persistent iptables-persistent"
+	echo "sudo rm -rf /run/live/medium/live/initrd.img*"
+	#TODO:ntps-juttui mukaan koska tar-nalkutus päiväyksistä
+	echo "sudo apt-get --no-install-recommends reinstall dnsmasq-base runit-helper"
+	echo "sudo rm -rf /run/live/medium/live/initrd.img*"
 
-echo "sudo apt-get --no-install-recommends reinstall libgetdns10 libbsd0 libidn2-0 libssl1.1 libunbound8 libyaml-0-2 stubby"
-echo "sudo rm -rf /run/live/medium/live/initrd.img*"
+	echo "sudo apt-get --no-install-recommends reinstall libgetdns10 libbsd0 libidn2-0 libssl1.1 libunbound8 libyaml-0-2 stubby"
+	echo "sudo rm -rf /run/live/medium/live/initrd.img*"
 
-echo "sudo tar -rvpf /mnt/186/nu.tar /var/cache/apt/archives/*.deb"
-echo "sudo /sbin/ifdown eth0"
+	#VAIH:mja
+	echo "sudo tar -rvpf ${tgtfile}  /var/cache/apt/archives/*.deb"
+	echo "sudo /sbin/ifdown ${iface}"
+else
+	echo "not fetching pkgs"
+fi
+
+#exit
 
 echo "sleep 5"
 sudo dpkg -i /var/cache/apt/archives/dns-root-data*.deb ; sudo rm -rf /var/cache/apt/archives/dns-root-data*.deb
@@ -90,24 +128,27 @@ sleep 2
 
 #sudo /etc/init.d/netfilter-persistent restart #VARMEMPI TOISELLA TAVALLA PRKL
 sudo /usr/sbin/ip6tables-restore /etc/iptables/rules.v6
-sudo /usr/sbin/iptables-restore /etc/iptables/rules.v4.180624b 
+sudo /usr/sbin/iptables-restore /etc/iptables/${tblz4}.b 
 
-sudo /etc/init.d/dnsmasq start
+sudo /etc/init.d/dnsmasq restart
 sleep 3
+#exit
 
 #tämän kanssa oli vielä jotain pientä 
 sudo adduser --system stubby
 sleep 5
 sudo /etc/init.d/stubby start
 sleep 5
+#190624:tähän asti vissiin ok
 
-#TODO:varmista tämä kohdan toimivuus
+#VAIH:varmista tämä kohdan toimivuus
 #https://raw.githubusercontent.com/senescent777/project/main/sbin/dhclient-script.new
 
+#HUOM.190624:parempi jotta mv, Daedalus nalkutti linkityksestä
 sudo mv /sbin/dhclient-script /sbin/dhclient-script.OLD
 sudo mv /sbin/dhclient-script.new /sbin/dhclient-script
 
 sudo chmod 0555 /sbin/dhclient*
 sudo chown root:root /sbin/dhclient*
 
-echo "sudo /sbin/ifup eth0"
+echo "sudo /sbin/ifup ${iface}"
